requirements.txt
Pillow
rembg
requests
flask
onnxruntime
gunicorn